requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
google-genai>=1.0
httpx>=0.27
pytest>=8.0
